fix: version endpoint
